Changes:
New module for the last tab, e-mail configuration and send.

Layout created for the above tab.

Added functionality in preview to jump to specific relative row.

Added new tab for attachment configuration.

Need to design layout and add functionality to the above tab.

Need to add functionality to send the e-mails.