Changes:
New modules to seperate code for each tab and for common messages.

Improved error handling.

Added the field and logic to get receiver e-mail addresses.

Updated preview with attachments and receiver e-mail address.

Created the layout and logic for the attachments tab.

Created the layout for the final tab, most of the logic already exists in preview module to get the data and make replacements, so I will need to share the functions.